backend/pyproject.toml
[tool.poetry]
name = "movie-back"
version = "0.1.1"
description = "back project for socotec.io test"
authors = ["Adrien Montagu <adrienmontagu@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
django = ">=2.2"
djangorestframework = "*"
drf-yasg = "*"
django-environ = "*"
django-cors-headers = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
black = "*"
isort = "*"
mock = "*"
django-filter = "^2.4.0"
psycopg2-binary = "^2.8.6"
pytest-django = "^4.2.0"
mkdocs = "^1.2.2"
freezegun = "^1.1.0"
toml = "^0.10.2"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 95
include = '\.pyi?$'
force-exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | examples
  | migrations
  | grpc
)/
'''

[tool.isort]
profile = "black"
include_trailing_comma = true
line_length = 95
multi_line_output = 3
use_parentheses = true
known_third_party = ['mock']
skip_glob = ["**/migrations/*","**/grpc/*"]

[tool.poetry.scripts]
tests = "test_utils.load_tests:launch"